+++ React/foes-db/src/components/file-view-section/FileViewSection.jsx
@@ -1,18 +1,28 @@
+import { Button } from '@mui/material';
 import React from 'react';
+import { useSelector } from 'react-redux';
+import DownloadIcon from '@mui/icons-material/Download';
 
 const FileViewSection = (props) => {
 
+  const token = useSelector(state => state.auth.tokenId);
+
   const onEditHandler = () => {
     props.setIsEditing(true);
-  }
+  };
 
   const onDeleteHandler = () => {
-    fetch(`http://127.0.0.1:8000/api/deleteAttachment/${props.obj.id}`, {
-      method: 'DELETE'
-    })
-    const newAttachments = (props.attachments).filter((element, i) => !(i === props.index));
-    props.setAttachments(newAttachments);
-  }
+    if (window.confirm("Are you sure you want to delete this element?") == true) {
+      fetch(`http://127.0.0.1:8000/api/deleteAttachment/${props.obj.id}`, {
+        method: 'DELETE',
+        headers: {
+          Authorization: `Bearer ${token}`
+        }
+      });
+      const newAttachments = (props.attachments).filter((element, i) => !(element.id === props.attachmentId));
+      props.setAttachments(newAttachments);
+    }
+  };
 
   return (
     <div className="section">
@@ -26,8 +36,13 @@ const FileViewSection = (props) => {
         <div className="formInput">
           <label>file_name: {props.obj.file_name}</label>
         </div>
-        <button onClick={onEditHandler}>edit</button>
-        <button onClick={onDeleteHandler}>delete</button>
+        <a href={`http://127.0.0.1:8000/api/downloadAttachment/${props.attachmentId}`} download='myFile'>
+          <Button variant="contained" color='success' endIcon={<DownloadIcon />} onClick={() => { (<a href={`http://127.0.0.1:8000/api/downloadAttachment/${props.attachmentId}`} download='myFile'>Download</a>); }}>
+            Download
+          </Button>
+        </a>
+        <Button className='section-btn' variant='contained' onClick={onEditHandler}>edit</Button>
+        <Button className='section-btn' variant='outlined' color='error' onClick={onDeleteHandler}>delete</Button>
       </div>
     </div>
   );
