+++ React/foes-db/src/components/key-person-view-section/KeyPersonViewSection.jsx
@@ -1,18 +1,28 @@
-import React from 'react'
+import { Button } from '@mui/material';
+import React from 'react';
+import { useSelector } from 'react-redux';
+import './keyPersonViewSection.scss'
 
 const KeyPersonViewSection = (props) => {
 
+  const token = useSelector(state => state.auth.tokenId);
+
   const onEditHandler = () => {
     props.setIsEditing(true);
-  }
+  };
 
   const onDeleteHandler = () => {
-    fetch(`http://127.0.0.1:8000/api/deleteKeyContactPerson/${props.obj.id}`, {
-      method: 'DELETE'
-    })
-    const newKeyPersons = (props.keyPersons).filter((element, i) => !(i === props.index));
-    props.setKeyPersons(newKeyPersons);
-  }
+    if (window.confirm("Are you sure you want to delete this element?") == true) {
+      fetch(`http://127.0.0.1:8000/api/deleteKeyContactPerson/${props.obj.id}`, {
+        method: 'DELETE',
+        headers: {
+          Authorization: `Bearer ${token}`
+        }
+      });
+      const newKeyPersons = (props.keyPersons).filter((element, i) => !(i === props.index));
+      props.setKeyPersons(newKeyPersons);
+    }
+  };
 
   return (
     <div className="section">
@@ -26,11 +36,11 @@ const KeyPersonViewSection = (props) => {
         <div className="formInput">
           <label>Institution: {props.obj.institution}</label>
         </div>
-        <button onClick={onEditHandler}>edit</button>
-        <button onClick={onDeleteHandler}>delete</button>
+        <Button className='section-btn' variant="contained"  onClick={onEditHandler}>edit</Button>
+        <Button className='section-btn' variant="contained" color="error" onClick={onDeleteHandler}>delete</Button>
       </div>
     </div>
-  )
-}
+  );
+};
 
-export default KeyPersonViewSection
+export default KeyPersonViewSection;
